refactor: pull selectedPreset parameter into TEPattern
